split RequestSender from GroongaClient
