+++ src/main/scala/com/naokia/groonga4s/command/package.scala
@@ -5,7 +5,7 @@ package object command {
     def getQuery: String
   }
 
-  trait PostCommand extends Command{
+  trait CommandWithBody extends Command{
     def getBody: String
   }
